+++ src/pages/Inicio.jsx
@@ -5,6 +5,7 @@ import VidaBendecida from '../components/VidaBendecida'
 import Eventos from '../components/Eventos'
 import Testimonios from '../components/Testimonios'
 import useTituloPagina from '../hooks/useTituloPagina'
+import { secciones } from '../data/contenido'
 
 export default function Inicio() {
   useTituloPagina()
@@ -12,11 +13,11 @@ export default function Inicio() {
   return (
     <>
       <Hero />
-      <Biografia />
-      <Ministerios />
-      <VidaBendecida />
-      <Eventos />
-      <Testimonios />
+      {secciones.biografia && <Biografia />}
+      {secciones.ministerios && <Ministerios />}
+      {secciones.vida_bendecida && <VidaBendecida />}
+      {secciones.eventos && <Eventos />}
+      {secciones.testimonios && <Testimonios />}
     </>
   )
 }